Removing the sub-class listeners
The plan will be to break the IssueListener into smaller pieces and
then selectively turn on/off different listeners for a repository.
But mostly, I want to keep the differences between repositories to a minimum,
to avoid these differences.
